devel_requirements.txt
-r requirements.txt
Flask-Testing < 0.5
names < 0.4
